requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
sqlalchemy>=2.0,<3
bcrypt>=4.1,<5
cryptography>=43,<45
pydantic>=2.8,<3
python-multipart>=0.0.9,<1
python-docx>=1.1,<2
pypdf>=5.0,<6
python-dotenv>=1.0,<2
